* Restricted update/delete buttons and views to users that are the host of a room.

+++ discord/base/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
@@ -76,6 +77,9 @@ def createRoom(request):
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here.')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -89,6 +93,10 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here.')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
